added deadlockdetector (and marking the current core only jdk1.6+ :) hellbound trust system initial commit hellbound faction basic support l2j 2152 castle managment
git-svn-id: http://svn.l2jfree.com/l2jfree@4049 6ff00ba9-25cd-4355-987e-099c62bc2628

+++ trunk/l2jfree-core/src/main/java/com/l2jfree/gameserver/model/actor/instance/L2ClanHallManagerInstance.java
@@ -177,7 +177,7 @@ public class L2ClanHallManagerInstance extends L2MerchantInstance
                     else
                     {
                         html.setFile("data/html/clanHallManager/support"+getClanHall().getFunction(ClanHall.FUNC_SUPPORT).getLvl()+".htm");
-                        html.replace("%mp%", String.valueOf(getStatus().getCurrentMp()));
+                        html.replace("%mp%", String.valueOf((int)getStatus().getCurrentMp()));
                     }
                     sendHtmlMessage(player, html);
                 }
@@ -1284,7 +1284,7 @@ public class L2ClanHallManagerInstance extends L2MerchantInstance
                         {
                             NpcHtmlMessage html = new NpcHtmlMessage(1);
                             html.setFile("data/html/clanHallManager/support-no_mana.htm");
-                            html.replace("%mp%", String.valueOf(getStatus().getCurrentMp()));
+                            html.replace("%mp%", String.valueOf((int)getStatus().getCurrentMp()));
                             sendHtmlMessage(player, html);
                             return;
                         }
@@ -1332,7 +1332,7 @@ public class L2ClanHallManagerInstance extends L2MerchantInstance
                 if(getClanHall().getFunction(ClanHall.FUNC_SUPPORT).getLvl() == 0) 
                     return; 
                 html.setFile("data/html/clanHallManager/support" + getClanHall().getFunction(ClanHall.FUNC_SUPPORT).getLvl()+".htm"); 
-                html.replace("%mp%", String.valueOf(getStatus().getCurrentMp())); 
+                html.replace("%mp%", String.valueOf((int)getStatus().getCurrentMp())); 
                 sendHtmlMessage(player, html); 
             }
             else if (actualCommand.equalsIgnoreCase("goto"))
